Fix glitch with switching users from QuickSettings
Since keyguard and systemui were merged into the same process to save
memory, they share the same Looper and graphics context.  As a result,
there's no way to allow concurrent animation while keyguard inflates.
The workaround is to add a slight delay to allow the animation to finish.

Workaround for bug 11046339

Change-Id: I355ca4fb325312bae87d587acfb12b475ecdeba4